Fix coverage table generation of empty files #145
When a file had no coverage points (for example when the whole file was
ignored) Cover would fail to include that file in its coverage map at
all. This is because the flatten/merge step of coverage table generation
would erase the empty coverage row. These files are now added directly
back in.

In addition this commit better describes the coverage table merging
in comments (as I had a hard time remember what the code did) and also
add more verbose debugging logs.